CartPoleV1/requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
backcall==0.2.0
bcc==0.18.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
cloudpickle==2.2.0
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.6
distro==1.7.0
docker-compose==1.29.2
docutils==0.19
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.0
GDAL==3.4.1
gym==0.26.2
gym-notices==0.0.8
hidpidaemon==18.4.6
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iotop==0.6
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jupyter-console==6.4.4
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
more-itertools==8.10.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
netaddr==0.8.0
netifaces==0.11.0
notebook_shim==0.2.2
numpy==1.21.5
oauthlib==3.2.0
packaging==21.3
pandocfilters==1.5.0
paramiko==2.12.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pipenv==11.9.0
platformdirs==2.5.4
pop-transition==1.1.2
prometheus-client==0.15.0
prompt-toolkit==3.0.33
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydbus==0.6.0
Pygments==2.13.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
repolib==2.1.1
repoman==1.4.0
requests==2.25.1
s-tui==1.1.3
SecretStorage==3.3.1
Send2Trash==1.8.0
sessioninstaller==0.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
ssh-import-id==5.11
stack-data==0.6.1
swig==4.1.1
systemd-python==234
terminado==0.17.0
tinycss2==1.2.1
torch==1.13.0+cpu
torchaudio==0.13.0+cpu
torchvision==0.14.0+cpu
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufw==0.36.1
urllib3==1.26.5
urwid==2.1.2
vboxapi==1.0
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.59.0
widgetsnbextension==4.0.3
xdg==5
xkit==0.0.0
zipp==1.0.0
